- Providers' concurrency

+++ aware-core/src/main/java/com/aware/providers/Battery_Provider.java
@@ -140,34 +140,15 @@ public class Battery_Provider extends ContentProvider {
 					+ Battery_Charges.BATTERY_END + " integer default 0,"
 					+ Battery_Charges.END_TIMESTAMP + " real default 0" };
 
-	private static UriMatcher sUriMatcher = null;
-	private static HashMap<String, String> batteryProjectionMap = null;
-	private static HashMap<String, String> batteryDischargesMap = null;
-	private static HashMap<String, String> batteryChargesMap = null;
-	private static DatabaseHelper databaseHelper = null;
-	private static SQLiteDatabase database = null;
+	private UriMatcher sUriMatcher = null;
+	private HashMap<String, String> batteryProjectionMap = null;
+	private HashMap<String, String> batteryDischargesMap = null;
+	private HashMap<String, String> batteryChargesMap = null;
+	private DatabaseHelper databaseHelper = null;
 
-	private boolean initializeDB() {
-        if (databaseHelper == null) {
-            databaseHelper = new DatabaseHelper( getContext(), DATABASE_NAME, null, DATABASE_VERSION, DATABASE_TABLES, TABLES_FIELDS );
-        }
-        if( databaseHelper != null && ( database == null || ! database.isOpen() )) {
-            database = databaseHelper.getWritableDatabase();
-        }
-        return( database != null && databaseHelper != null);
-    }
-
-	/**
-	 * Recreates the ContentProvider
-	 */
-	public static void resetDB( Context c ) {
-		Log.d("AWARE", "Resetting " + DATABASE_NAME + "...");
-
-		File db = new File(DATABASE_NAME);
-		db.delete();
-		databaseHelper = new DatabaseHelper( c, DATABASE_NAME, null, DATABASE_VERSION, DATABASE_TABLES, TABLES_FIELDS);
-		if( databaseHelper != null ) {
-			database = databaseHelper.getWritableDatabase();
+	private void initializeDB() {
+		if (databaseHelper == null) {
+			databaseHelper = new DatabaseHelper(getContext(), DATABASE_NAME, null, DATABASE_VERSION, DATABASE_TABLES, TABLES_FIELDS);
 		}
 	}
 	
@@ -177,37 +158,35 @@ public class Battery_Provider extends ContentProvider {
 	@Override
 	public int delete(Uri uri, String selection, String[] selectionArgs) {
 
-	    if( ! initializeDB() ) {
-            Log.w(AUTHORITY,"Database unavailable...");
-            return 0;
-        }
+		initializeDB();
 
-		int count = 0;
+		SQLiteDatabase database = databaseHelper.getWritableDatabase();
+		if (database == null) return 0;
+
+		//lock database for transaction
+		database.beginTransaction();
+
+		int count;
 		switch (sUriMatcher.match(uri)) {
 		case BATTERY:
-            database.beginTransaction();
 			count = database.delete(DATABASE_TABLES[0], selection,
 					selectionArgs);
-            database.setTransactionSuccessful();
-            database.endTransaction();
 			break;
 		case BATTERY_DISCHARGE:
-            database.beginTransaction();
 			count = database.delete(DATABASE_TABLES[1], selection,
 					selectionArgs);
-            database.setTransactionSuccessful();
-            database.endTransaction();
 			break;
 		case BATTERY_CHARGE:
-            database.beginTransaction();
 			count = database.delete(DATABASE_TABLES[2], selection,
 					selectionArgs);
-            database.setTransactionSuccessful();
-            database.endTransaction();
 			break;
 		default:
+			database.endTransaction();
 			throw new IllegalArgumentException("Unknown URI " + uri);
 		}
+
+		database.setTransactionSuccessful();
+		database.endTransaction();
 
 		getContext().getContentResolver().notifyChange(uri, null);
 		return count;
@@ -239,17 +218,17 @@ public class Battery_Provider extends ContentProvider {
 	@Override
 	public Uri insert(Uri uri, ContentValues initialValues) {
 
-	    if( ! initializeDB() ) {
-            Log.w(AUTHORITY,"Database unavailable...");
-            return null;
-        }
+		initializeDB();
 
-		ContentValues values = (initialValues != null) ? new ContentValues(
-				initialValues) : new ContentValues();
+		SQLiteDatabase database = databaseHelper.getWritableDatabase();
+		if (database == null) return null;
+
+		ContentValues values = (initialValues != null) ? new ContentValues(initialValues) : new ContentValues();
+
+		database.beginTransaction();
 
 		switch (sUriMatcher.match(uri)) {
 		case BATTERY:
-            database.beginTransaction();
 			long battery_id = database.insertWithOnConflict(DATABASE_TABLES[0], Battery_Data.TECHNOLOGY, values, SQLiteDatabase.CONFLICT_IGNORE);
             database.setTransactionSuccessful();
             database.endTransaction();
@@ -258,9 +237,9 @@ public class Battery_Provider extends ContentProvider {
                 getContext().getContentResolver().notifyChange(batteryUri, null);
                 return batteryUri;
             }
+			database.endTransaction();
             throw new SQLException("Failed to insert row into " + uri);
 		case BATTERY_DISCHARGE:
-            database.beginTransaction();
 			long battery_d_id = database.insertWithOnConflict(DATABASE_TABLES[1], Battery_Discharges.DEVICE_ID, values, SQLiteDatabase.CONFLICT_IGNORE);
             database.setTransactionSuccessful();
             database.endTransaction();
@@ -271,9 +250,9 @@ public class Battery_Provider extends ContentProvider {
                         .notifyChange(batteryUri, null);
                 return batteryUri;
             }
+			database.endTransaction();
             throw new SQLException("Failed to insert row into " + uri);
 		case BATTERY_CHARGE:
-            database.beginTransaction();
             long battery_c_id = database.insertWithOnConflict(DATABASE_TABLES[2],
                     Battery_Charges.DEVICE_ID, values, SQLiteDatabase.CONFLICT_IGNORE);
             database.setTransactionSuccessful();
@@ -285,8 +264,10 @@ public class Battery_Provider extends ContentProvider {
                         .notifyChange(batteryUri, null);
                 return batteryUri;
             }
+			database.endTransaction();
             throw new SQLException("Failed to insert row into " + uri);
 		default:
+			database.endTransaction();
 			throw new IllegalArgumentException("Unknown URI " + uri);
 		}
 	}
@@ -364,10 +345,10 @@ public class Battery_Provider extends ContentProvider {
 	public Cursor query(Uri uri, String[] projection, String selection,
 			String[] selectionArgs, String sortOrder) {
 
-	    if( ! initializeDB() ) {
-            Log.w(AUTHORITY,"Database unavailable...");
-            return null;
-        }
+		initializeDB();
+
+		SQLiteDatabase database = databaseHelper.getReadableDatabase();
+		if (database == null) return null;
 
 		SQLiteQueryBuilder qb = new SQLiteQueryBuilder();
 		switch (sUriMatcher.match(uri)) {
@@ -405,37 +386,34 @@ public class Battery_Provider extends ContentProvider {
 	public int update(Uri uri, ContentValues values, String selection,
 			String[] selectionArgs) {
 
-	    if( ! initializeDB() ) {
-            Log.w(AUTHORITY,"Database unavailable...");
-            return 0;
-        }
+		initializeDB();
+
+		SQLiteDatabase database = databaseHelper.getWritableDatabase();
+		if (database == null) return 0;
+
+		database.beginTransaction();
 
 		int count;
 		switch (sUriMatcher.match(uri)) {
 		case BATTERY:
-            database.beginTransaction();
 			count = database.update(DATABASE_TABLES[0], values, selection,
 					selectionArgs);
-            database.setTransactionSuccessful();
-            database.endTransaction();
 			break;
 		case BATTERY_DISCHARGE:
-            database.beginTransaction();
 			count = database.update(DATABASE_TABLES[1], values, selection,
 					selectionArgs);
-            database.setTransactionSuccessful();
-            database.endTransaction();
 			break;
 		case BATTERY_CHARGE:
-            database.beginTransaction();
 			count = database.update(DATABASE_TABLES[2], values, selection,
 					selectionArgs);
-            database.setTransactionSuccessful();
-            database.endTransaction();
 			break;
 		default:
+			database.endTransaction();
 			throw new IllegalArgumentException("Unknown URI " + uri);
 		}
+
+		database.setTransactionSuccessful();
+		database.endTransaction();
 
 		getContext().getContentResolver().notifyChange(uri, null);
 		return count;
